refactor: useProductDetail 리팩터링(파일분리)
